hotswap: split loader test support
